# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Agentic Coder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files/tree:
    post:
      operationId: getFileTree
      tags: [files]
      summary: Get directory tree structure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileTreeRequest"
      responses:
        "200":
          description: Directory tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileTreeNode"
  /files/read:
    post:
      operationId: readFile
      tags: [files]
      summary: Read file content (with optional chunking)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileReadRequest"
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileReadResult"
  /files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write or update file content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWriteRequest"
      responses:
        "200":
          description: Write result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileOpResult"
  /files/delete:
    post:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileDeleteRequest"
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileOpResult"
  /files/mkdir:
    post:
      operationId: createDirectory
      tags: [files]
      summary: Create a directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MkdirRequest"
      responses:
        "200":
          description: Mkdir result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileOpResult"
  /files/patch:
    post:
      operationId: patchFile
      tags: [files]
      summary: Apply a targeted patch (search and replace) to a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilePatchRequest"
      responses:
        "200":
          description: Patch result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileOpResult"
  /files/search:
    post:
      operationId: searchInFiles
      tags: [files]
      summary: Search for text across files in a directory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileSearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileSearchResult"
  /execute:
    post:
      operationId: executeCommand
      tags: [execute]
      summary: Execute a shell command and get output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteRequest"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteResult"
...
  /execute/sessions/{terminalId}/kill:
    post:
      operationId: killTerminalSession
      tags: [execute]
      summary: Kill a running terminal session
      parameters:
        - name: terminalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Kill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileOpResult"
...
components:
  schemas:
...
    FileTreeRequest:
      type: object
      properties:
        path:
          type: string
        depth:
          type: integer
        ignorePatterns:
          type: array
          items:
            type: string
      required:
        - path
    FileTreeNode:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
        size:
          type: ["integer", "null"]
        children:
          type: array
          items:
            $ref: "#/components/schemas/FileTreeNode"
      required:
        - name
        - path
        - type
    FileReadRequest:
      type: object
      properties:
        path:
          type: string
        startLine:
          type: integer
        endLine:
          type: integer
        encoding:
          type: string
      required:
        - path
    FileReadResult:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        totalLines:
          type: integer
        startLine:
          type: integer
        endLine:
          type: integer
        truncated:
          type: boolean
        size:
          type: integer
      required:
        - path
        - content
        - totalLines
        - startLine
        - endLine
        - truncated
        - size
    FileWriteRequest:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        createDirs:
          type: boolean
        backup:
          type: boolean
      required:
        - path
        - content
    FileDeleteRequest:
      type: object
      properties:
        path:
          type: string
        recursive:
          type: boolean
      required:
        - path
    MkdirRequest:
      type: object
      properties:
        path:
          type: string
        recursive:
          type: boolean
      required:
        - path
    FilePatchRequest:
      type: object
      properties:
        path:
          type: string
        patches:
          type: array
          items:
            $ref: "#/components/schemas/FilePatch"
      required:
        - path
        - patches
    FilePatch:
      type: object
      properties:
        search:
          type: string
        replace:
          type: string
        replaceAll:
          type: boolean
      required:
        - search
        - replace
    FileOpResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        path:
          type: string
      required:
        - success
        - message
    FileSearchRequest:
      type: object
      properties:
        rootPath:
          type: string
        query:
          type: string
        filePattern:
          type: string
        maxResults:
          type: integer
        caseSensitive:
          type: boolean
      required:
        - rootPath
        - query
    FileSearchResult:
      type: object
      properties:
        file:
          type: string
        line:
          type: integer
        content:
          type: string
        context:
          type: string
      required:
        - file
        - line
        - content
    ExecuteRequest:
      type: object
      properties:
        command:
          type: string
        cwd:
          type: string
        timeout:
          type: integer
        sessionId:
          type: string
      required:
        - command
    ExecuteResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        duration:
          type: integer
        timedOut:
          type: boolean
        sessionId:
          type: string
      required:
        - stdout
        - stderr
        - exitCode
        - duration
        - timedOut
...