# lib/api-zod/src/generated/types/executeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Agentic Coder API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteRequest {
  command: string;
  cwd?: string;
  timeout?: number;
  sessionId?: string;
}