# lib/api-zod/src/generated/types/filePatchRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Agentic Coder API
 * OpenAPI spec version: 0.1.0
 */
import type { FilePatch } from './filePatch';

export interface FilePatchRequest {
  path: string;
  patches: FilePatch[];
}